# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CRMAI - Full Featured CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List opportunities
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: accountId
          in: query
          schema:
            type: integer
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: List of opportunities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityList"
    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Create opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpportunityInput"
      responses:
        "201":
          description: Created opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      summary: Get opportunity by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOpportunityInput"
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    delete:
      operationId: deleteOpportunity
      tags: [opportunities]
      summary: Delete opportunity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /opportunities/{id}/items:
    get:
      operationId: listOpportunityItems
      tags: [opportunities]
      summary: List opportunity line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of opportunity items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OpportunityItem"
    put:
      operationId: updateOpportunityItems
      tags: [opportunities]
      summary: Replace opportunity line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/CreateQuoteItemInput"
      responses:
        "200":
          description: Updated opportunity items
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OpportunityItem"

  # ACTIVITIES
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: integer
      required:
        - success
        - id

    # USER
...
    Opportunity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        accountId:
          type: integer
          nullable: true
        accountName:
          type: string
          nullable: true
        contactId:
          type: integer
          nullable: true
        contactName:
          type: string
          nullable: true
        stage:
          type: string
          enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
        amount:
          type: number
          nullable: true
        probability:
          type: integer
          nullable: true
        closeDate:
          type: string
          format: date-time
          nullable: true
        description:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        leadSource:
          type: string
          nullable: true
        nextStep:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - stage
        - createdAt
        - updatedAt
    OpportunityList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateOpportunityInput:
      type: object
      properties:
        name:
          type: string
        accountId:
          type: integer
          nullable: true
        contactId:
          type: integer
          nullable: true
        priceBookId:
          type: integer
          nullable: true
        stage:
          type: string
          enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
          default: prospecting
        amount:
          type: number
          nullable: true
        probability:
          type: integer
          nullable: true
        closeDate:
          type: string
          format: date-time
          nullable: true
        description:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        leadSource:
          type: string
          nullable: true
        nextStep:
          type: string
          nullable: true
      required:
        - name
    UpdateOpportunityInput:
      type: object
      properties:
        name:
          type: string
        accountId:
          type: integer
          nullable: true
        contactId:
          type: integer
          nullable: true
        priceBookId:
          type: integer
          nullable: true
        stage:
          type: string
          enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
        amount:
          type: number
          nullable: true
        probability:
          type: integer
          nullable: true
        closeDate:
          type: string
          format: date-time
          nullable: true
        description:
          type: string
          nullable: true
        assignedTo:
          type: integer
          nullable: true
        leadSource:
          type: string
          nullable: true
        nextStep:
          type: string
          nullable: true

    # ACTIVITY
...
    OpportunityItem:
      type: object
      properties:
        id:
          type: integer
        opportunityId:
          type: integer
        productId:
          type: integer
          nullable: true
        priceBookEntryId:
          type: integer
          nullable: true
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        discount:
          type: number
        total:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - opportunityId
        - productName
        - quantity
        - unitPrice
        - discount
        - total
    CreateQuoteItemInput:
      type: object
      properties:
        productId:
          type: integer
          nullable: true
        priceBookEntryId:
          type: integer
          nullable: true
        productName:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        discount:
          type: number
          default: 0
      required:
        - productName
        - quantity
        - unitPrice
...